requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.3
configobj==5.0.6
DateTime==4.3
Django==3.0.1
idna==2.8
ipython==7.10.1
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.4.2
mccabe==0.6.1
memory-profiler==0.55.0
nibabel==2.5.1
numpy==1.17.4
pandas==0.25.3
prompt-toolkit==3.0.2
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
requests==2.22.0
six==1.13.0
soupsieve==1.9.5
Sphinx==2.2.2
sphinx-bootstrap-theme==0.7.1
sphinx-fontawesome==0.0.6
sphinx-gallery==0.6.0.dev0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
stdlib-list==0.6.0
typed-ast==1.4.0
urllib3==1.25.7
vtk==8.1.2
wincertstore==0.2
zope.interface==4.7.1
